[Ref]: 스트리밍 플랫폼 및 공유 기능 코드 리팩토링
- `PlatformInfoSheetViewModel` 추가
  - `PlatformInfoSheetView`가 직접 데이터를 관리하는 대신 ViewModel을 통해 관리
  - 영화 ID 기반으로 스트리밍 플랫폼 목록을 로드하는 `loadPlatforms()` 구현

- `SharingOptionsViewModel` 추가
  - 공유 관련 로직을 `ViewModel`로 분리하여 관리
  - `copyURL()`, `shareWithKakaoTalk()`, `sendMessage()`, `systemShare()` 등의 함수 추가
  - `showToastWith(message:)` 함수를 통해 모든 공유 방식에서 토스트 메시지를 쉽게 표시

- `DetailActionButtonsView` 수정
  - 공유 버튼에 기존 `showCustomShareSheet` 상태값을 제거하고 `viewModel.showShareSheet`로 변경
  - `PlatformInfoSheetView`가 `.presentationDetents([.large])`을 사용해 UI 개선

- `PlatformInfoSheetView` 개선
  - 기존의 `AvailablePlatformsView`, `EmptyPlatformsView`, `RentalPurchaseInfoView`를 모듈화하여 코드 정리
  - `PlatformInfoSheetTheme`를 별도 파일로 분리하여 UI 색상 및 스타일 일관성 유지

- `CustomShareSheetView` 개선
  - `SharingOptionsListView`를 내부에 포함하여 가독성 향상
  - `MoviePreviewCard` 추가하여 공유 시 영화 정보를 미리 볼 수 있도록 개선

- `MoviePreviewCard` 개선
  - 영화 제목, 장르, 개봉일, 평점 등을 한눈에 볼 수 있도록 UI 정리
  - `Theme` 내부에서 텍스트 및 배경 스타일 정의하여 일관성 유지

- `RentalPurchaseInfoView` 추가
  - 기존 `PlatformInfoSheetView` 내부에 있던 대여 및 구매 정보를 별도 뷰로 분리하여 재사용성 증가

- `SharingOptionsListView` 개선
  - `SharingOptionsViewModel`을 사용하여 공유 로직 분리
  - 메시지 공유, URL 복사, 카카오톡 공유 등 이벤트 처리 로직 정리

- `Podfile` 업데이트
  - `KakaoSDKCommon`, `KakaoSDKAuth`, `KakaoSDKUser`, `KakaoSDKShare` 추가하여 카카오톡 공유 기능 지원
